Updated colour scheme
changed colours from blue to green (matter of preference). changed the photo at the top of the page to an image of my C++ knights tale game as this required better art skills to produce compared to the stars. Updated the stylesheet in multiple other areas to make the website look better overall on both mobile and larger screens.